+++ Hazel/Src/Hazel/Renderer/RenderCommandRegister.h
@@ -27,11 +27,11 @@ namespace Hazel
 
 		// 添加各种类型的DrawQuad函数, 包含了position、rotation、texture、tiling和tintColor
 
-		static void DrawSpriteRenderer(const SpriteRenderer& spriteRenderer, const glm::mat4& transform, uint32_t goId);
+		static void DrawSpriteRenderer(const SpriteRenderer& spriteRenderer, const glm::mat4& transform, uint64_t goId);
 
-		static void DrawQuad(uint32_t goId, const glm::mat4& transform, const glm::vec4& color);
-		static void DrawQuad(uint32_t goId, const glm::mat4& transform, const std::shared_ptr<Texture2D>& texture, float tilingFactor = 1.0f, const glm::vec4& tintColor = { 1,1,1,1 });
-		static void DrawQuad(uint32_t goId, const glm::mat4& transform, const std::shared_ptr<SubTexture2D>& subTexture, float tilingFactor = 1.0f, const glm::vec4& tintColor = { 1,1,1,1 });
+		static void DrawQuad(uint64_t goId, const glm::mat4& transform, const glm::vec4& color);
+		static void DrawQuad(uint64_t goId, const glm::mat4& transform, const std::shared_ptr<Texture2D>& texture, float tilingFactor = 1.0f, const glm::vec4& tintColor = { 1,1,1,1 });
+		static void DrawQuad(uint64_t goId, const glm::mat4& transform, const std::shared_ptr<SubTexture2D>& subTexture, float tilingFactor = 1.0f, const glm::vec4& tintColor = { 1,1,1,1 });
 
 		static std::shared_ptr<Shader> GetCurrentShader();
